Updating README and moved Image component to a better folder next to Main component
