+++ README.md
@@ -10,6 +10,8 @@ Check out the live site here: https://wondrous-fudge-759bc2.netlify.app/
 
 ## Designed with 
 
-- React Props and Hooks 
+- React useEffect, useState, useContext Hooks 
+- Use of parent components where appropriate to avoid prop drilling
+= Reusable component structure
 - CSS 
 - Axios calls to the Dnd 5e RESTFUL API (http://www.dnd5eapi.co/docs/#get-/api/races/-index-)
